Utils: move the Split helpers out of convar.h into a separate header
Splitting a string on a set of delimiters is generally useful,
and is used in several places where pulling in convar.h isn't
warranted (in particular Utils/Process.cpp).
